Renamed NotInitialized  to comform with the language standard
