+++ include/NovelRT/Exceptions/NotInitialisedException.h
